capturing_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
cryptography
python-dotenv
numpy
httpx
insightface
onnxruntime
opencv-python
Pillow
psycopg2-binary
bcrypt
